backend/requirements.txt
fastapi
uvicorn[standard]

sqlalchemy
psycopg2-binary
pydantic-settings

# Testing
pytest
httpx

# Development
python-dotenv
